Refactor privates that share a time slot to use a wrapper lesson.
Previously, privates that needed to share a time block with one another
had a complex system of pointing to each other, which was inefficient
and difficult to reason about.

By creating a wrapper lesson to contain the shared privates it becomes
much easier to reason about which lessons are available to share a
block.

Closes #6